% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{BayesBoot}
\alias{BayesBoot}
\title{Internal helper function}
\usage{
BayesBoot(df, J)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
