% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{fit_model}
\alias{fit_model}
\title{Internal helper function}
\usage{
fit_model(X, Y, is_binary, opts, hypers, suppress)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
