% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{generate_random_matrix}
\alias{generate_random_matrix}
\title{Internal helper function}
\usage{
generate_random_matrix(J, regime, param)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
