% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{get_lagged_X}
\alias{get_lagged_X}
\title{Internal helper function}
\usage{
get_lagged_X(data, var.type, i, id, Lag1)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
