% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYhat_bart.R
\name{getYhat_bart}
\alias{getYhat_bart}
\title{Map fitted latents to categorical prediction}
\usage{
getYhat_bart(j, ndim, n, releveled, maxy, mu, sigmas, alpha)
}
\arguments{
\item{j}{Index of the current round of posterior draws,}

\item{ndim}{Number of latents,}

\item{n}{Sample size,}

\item{releveled}{The index-category dictionary returned from multinomial probit BART fit,}

\item{maxy}{The index of the reference level from within the multinomial probit BART core,}

\item{mu}{Sum of trees fit in 2 dimensions - (n x ndraws) x ndim,}

\item{sigmas}{Posteriors of the covariance matrix in 2 dimensions - (ndim x ndim) x ndraws; the jth column is the covariance matrix estimated from jth round of posterior draw by column,}
}
\value{
yhat Simulated outcome as a vector of length n,
}
\description{
Map fitted latents to categorical prediction for the jth posterior draw,
}
