% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{handle_D}
\alias{handle_D}
\title{Internal helper function}
\usage{
handle_D(
  x,
  j,
  BModels,
  iterations,
  it,
  Lag1,
  tgroup,
  b_size,
  drop,
  d,
  drop_param
)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
