% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{handle_Fi}
\alias{handle_Fi}
\title{Internal helper function}
\usage{
handle_Fi(x, fixed.regime, l)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
