% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{handle_Ra}
\alias{handle_Ra}
\title{Internal helper function}
\usage{
handle_Ra(
  x,
  j,
  continuous,
  BModels,
  iterations,
  it,
  random.regime,
  o,
  param,
  above,
  threshold,
  nat_value,
  incremental,
  Lag1,
  tgroup,
  b_size
)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
