% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{handle_S}
\alias{handle_S}
\title{Internal helper function}
\usage{
handle_S(
  x,
  j,
  BModels,
  iterations,
  it,
  Lag1,
  tgroup,
  b_size,
  s_hat,
  n_Reg,
  s,
  ks,
  drop
)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
