% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{handle_X}
\alias{handle_X}
\title{Internal helper function}
\usage{
handle_X(x, j, continuous, BModels, iterations, it, Lag1, tgroup, b_size)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
