% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{handle_Y}
\alias{handle_Y}
\title{Internal helper function}
\usage{
handle_Y(
  x,
  j,
  continuous,
  BModels,
  iterations,
  it,
  Lag1,
  tgroup,
  b_size,
  drop_shift,
  d,
  y,
  ky,
  n_Reg,
  Lag1_flag
)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
