% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{make_hypers}
\alias{make_hypers}
\title{Internal helper function}
\usage{
make_hypers(X, Y, tgroup, num_tree, alpha, eta, phi, alpha_vec, alpha_shape_1)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
