% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSB.R
\name{MakeForest}
\alias{MakeForest}
\title{MakeForest}
\usage{
MakeForest(hypers, opts)
}
\arguments{
\item{hypers}{List of hyperparameters.}

\item{opts}{List of options.}
}
\value{
An object of class Forest.
}
\description{
Creates a Forest object from Rcpp module.
}
\examples{
\dontrun{
X <- matrix(rnorm(100), nrow = 10)
Y <- rnorm(10)
hypers <- Hypers(X = X, Y = Y)
opts <- Opts()
forest <- MakeForest(hypers, opts)
}
}
