% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bart.R
\name{model_bart}
\alias{model_bart}
\title{Bayesian Additive Regression Trees}
\usage{
model_bart(formula = NULL, Yname = NULL, Xname = NULL, data, type,
  base = NULL, Prior = NULL, Mcmc = NULL, diagnostics = TRUE,
  make.prediction = TRUE, correction = TRUE, fitMNP = NULL,
  bylatent = NULL, Kindo = FALSE)
}
\arguments{
\item{formula}{response ~ covariates,}

\item{Yname}{Name of response,}

\item{Xname}{List of covariate names, with the ith element being the vector of covariate names for the ith latent variable; the order of the latent variables is in the numeric order of response categories after excluding the reference level,}

\item{data}{Training Data with the response and covariates in formula or Xname,}

\item{type}{Type of response, options are "continuous", "binary", and "multinomial",}

\item{base}{In the case of multinomial response, sets the reference level of the multinomial response. For example, if the response takes values 2, 3, and 4, then base = "3" sets response value 3 as the reference. Default is the highest class,}

\item{Prior}{List of Priors for continuous BART: e.g., Prior = list(nu=3,sigq=0.9, ntrees=200,  kfac=2.0, pswap=0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, minobsnode = 10). List of Priors for Probit BART: e.g., Prior = list(ntrees=200,  kfac=2.0, pswap=0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, minobsnode = 10). List of Priors for multinomial Probit BART: e.g., Prior = list(nu=p+2,  V= diag(p - 1), ntrees=200,  kfac=2.0, pswap=0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, minobsnode = 10).
The components of Prior are
\itemize{
\item nu: For continuous response: the degree of freedom in the inverse chi-square prior distribution of the error variance; for multinomial response: The covariance matrix of latent variables is assumed to have prior distribution Inv-Wish(nu, V), nu is the degree of freedom and nu > (nlatent - 1).
\item sigq: In the case of continuous response, the quantile of the error variance prior that the rough estimate (from linear regression) is placed at.
\item V : In the case of multinomial response, the positive definite scale matrix in the Inverse-Wishart prior of the covariance matrix.
\item ntrees : The total number of trees in each round of BART fitting, or a vector with the ith element being the number of trees for the ith latent variable when Yname and Xname are used,
\item kfac : A tuning parameter that satisfies mu - kfac * sigma = ymin and mu + kfac * sigma = ymax, where mu and sigma are the mean and std of the Gaussian prior distribution of the sum of fit of all trees.
\item pswap : The prior probability of swap move in simulating trees; default 0, there should be pswap no larger than 1-pbd.
\item pbd : The prior probability of birth/death move in simulating trees; default 1.
\item pb : The prior probability of birth given birth/death; default 0.5.
\item alpha : The prior probability of a bottom node splits is alpha/(1+d)^beta, d is depth of node.
\item beta : see alpha.
\item nc : The number of equally spaced cutpoints between min and max for each covariate.
\item minobsnode : The minimum number of observations in bottom nodes for birth in simulating trees.
}}

\item{Mcmc}{List of MCMC starting values, burn-in ... for continuous or binary response: e.g., list(burn = 100, ndraws = 1000); for multinomial response: e.g.  list(sigma0 = diag(p - 1), burn = 100, ndraws = 1000, nSigDr = 50)
#'The components of Mcmc are
\itemize{
\item w0 : The n x nlatent matrix of starting values for the latent variables; this can be results from the MNP. 
\item sigma0 : The nlatent x nlatent matrix of starting value for the covariance matrix of latent variables.
\item nSigDr: User-specified upper limit to repeated draws of the covariance variance matrix of latent variables in each round of posterior draw when condition 10 in Jiao and van Dyk 2015 is not satisfied. Default 50. 
}}

\item{diagnostics}{Returns convergence diagnostics and variable inclusion proportions if True (default),}

\item{make.prediction}{Returns simulated outcome samp_y if TRUE (default); FALSE is only applicable to continuous and binary outcomes,}

\item{correction}{Analysis applies correction from Jiao and van Dyk (2015) if TRUE (default); framework from Burgette and Nordheim (2012) is used if FALSE,}

\item{fitMNP}{If type is multinomial, then fitMNP is the number of rounds to pre-fit sum-of-trees on the w0 conditional on sigma0, default is zero; if fitMNP is not zero, there must be input for Mcmc w0 and sigma0,}

\item{bylatent}{If 0 then trees are updated across latents in the order of trees, otherwise across all trees in the order of latents,}

\item{Kindo}{Implement the modified version of MPBART by Kindo et al (2016) if TRUE; our modified MPBART with accelerated convergence is applied if FALSE (default),}
}
\value{
treefit ndraws x n posterior matrix of the training data sum of trees fit,

samp_y ndraws x n posterior matrix of the simulated outcome,

sigmasample posterior samples of the error standard deviation.

Percent_Acceptance Percent acceptance of Metropolis-Hastings proposals across the ntrees number of trees for each posterior draw after burn-in,

Tree_Num_Nodes Average number of tree nodes across the ntrees number of trees for each posterior draw after burn-in,

Tree_Num_Leaves Average number of leaves across the ntrees number of trees for each posterior draw after burn-in,

Tree_Depth Average tree depth across the ntrees number of trees for each posterior draw after burn-in,

Inclusion_Proportions Predictor inclusion frequencies. Smaller value of ntrees (such as 10, 20, 50, 100) is recommended for the purposes of variable selection.

TreeMod Object of the estimated sum of trees model; Treemod and sigmasample are needed for reusing the estimated model for predictions.
}
\description{
Bayesian Additive Regression Trees Modeling for Continuous, Binary, or Multinomial Outcome,
}
\examples{
##simulate data (example from Friedman MARS paper)
f = function(x){
 10*sin(pi*x[,1]*x[,2]) + 20*(x[,3]-.5)^2+10*x[,4]+5*x[,5]
}
sigma = 1.0 #y = f(x) + sigma*z , z~N(0,1)
n = 100 #number of observations
set.seed(99)
x=matrix(runif(n*10),n,10) #10 variables, only first 5 matter
Ey = f(x)
y=Ey+sigma*rnorm(n)
dat = data.frame(x,y)
fml = y ~ X1+X2+X3+X4+X5+X6+X7+X8+X9+X10
bfit = model_bart(fml, data = dat, type = "continuous",
                Prior = list(nu = 3, sigq = 0.9,
                             ntrees = 100,
                             kfac = 2,
                             pswap = 0.1, pbd = 0.5, pb = 0.25,
                             alpha = 0.95, beta = 2.0,
                             nc = 100, minobsnode = 10),
                Mcmc = list(burn=100, ndraws = 1000))

}
