% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{phi}
\alias{phi}
\title{Internal helper function}
\usage{
phi(a, b)
}
\arguments{
\item{a}{Numeric. Survival probability under intervention.}

\item{b}{Numeric. Survival probability under natural course.}
}
\value{
A numeric value equal to \code{max(1, a/b)}.
}
\description{
Compute Ratio Adjustment for Survival Probabilities.
Returns the maximum of 1 and the ratio \code{a/b}. Used in the computation of the SACE adjustment term.
}
\keyword{internal}
