% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{pred_comb}
\alias{pred_comb}
\title{Internal helper function}
\usage{
pred_comb(cont, BM, MCdatatmp, IT, L1, TG, bsize)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
