% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{pred_drop}
\alias{pred_drop}
\title{Internal helper function}
\usage{
pred_drop(BM, MCdata, IT, L1, TG, bsize)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
