% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{pred_ra}
\alias{pred_ra}
\title{Internal helper function}
\usage{
pred_ra(
  cont,
  BM,
  MCdatatmp,
  IT,
  random.regime,
  param,
  above,
  cutoff,
  nat_value,
  incremental,
  L1,
  TG,
  bsize
)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
