% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bart.R
\name{predict_bart}
\alias{predict_bart}
\title{Predict Method for BART Fits}
\usage{
predict_bart(obj, newdata = NULL, Gcomp = FALSE)
}
\arguments{
\item{obj}{Fitted model object from BART_mod,}

\item{newdata}{The data matrix to look for variables with which to predict, typically without the response; if Gcomp equals TRUE, number of rows should equal n x ndraws, where n is the number of subject. For the ith subject, (i, i+n, ..., i+(ndraws-1)n) rows are its simulated posterior outcomes from the previous simulation,}

\item{Gcomp}{Make predictions in the format of dynamic G-computation if TRUE; the default is FALSE,}
}
\value{
treefit ndraws x n posterior matrix of the sum of trees fit,

samp_y ndraws x n posterior matrix of the simulated outcome,
}
\description{
Predicted values based on Bayesian Additive Regression Trees model object,
}
\examples{
##simulate data (example from Friedman MARS paper)
f = function(x){
 10*sin(pi*x[,1]*x[,2]) + 20*(x[,3]-.5)^2+10*x[,4]+5*x[,5]
}
sigma = 1.0 #y = f(x) + sigma*z , z~N(0,1)
n = 100 #number of observations
set.seed(99)
x=matrix(runif(n*10),n,10) #10 variables, only first 5 matter
newx=matrix(runif(n*10),n,10) 
Ey = f(x)
y=Ey+sigma*rnorm(n)
dat = data.frame(x,y)
fml = as.formula("y ~ X1+X2+X3+X4+X5+X6+X7+X8+X9+X10")
bfit = BART_call(fml, data = dat, test.data = NULL,
                Prior = list(nu = 3, sigq = 0.9,
                             ntrees = 100,
                             kfac = 2,
                             pswap = 0.1, pbd = 0.5, pb = 0.25,
                             alpha = 0.95, beta = 2.0,
                             nc = 100, minobsnode = 10),
                Mcmc = list(burn=100, ndraws = 1000))

newy = predict_bart(bfit, newx)
}
