% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{prepare_Fi_X0_data}
\alias{prepare_Fi_X0_data}
\title{Internal helper function}
\usage{
prepare_Fi_X0_data(data, var.type, fixed.regime, J)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
