% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSB.R
\name{rlgam}
\alias{rlgam}
\title{rlgam}
\arguments{
\item{shape}{Shape parameter.}
}
\value{
A random gamma value.
}
\description{
Internal helper wrapping C++ for sampling from a gamma distribution.
}
\keyword{internal}
