% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{safe_val}
\alias{safe_val}
\title{Internal helper function}
\usage{
safe_val(x)
}
\description{
This function supports the main gcomputation workflow and is not intended for direct use.
}
\keyword{internal}
