% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSB.R
\name{update_sigma}
\alias{update_sigma}
\title{update_sigma}
\arguments{
\item{r}{Numeric value.}

\item{sigma_hat}{Estimated sigma.}

\item{sigma_old}{Previous sigma value.}

\item{temperature}{Temperature parameter for MH step.}
}
\value{
Updated sigma value.
}
\description{
Internal helper wrapping C++ for updating sigma in MCMC.
}
\keyword{internal}
