% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCSB.R
\name{sim_lfried}
\alias{sim_lfried}
\title{Simulate Data Using Modified Friedman Function}
\usage{
sim_lfried(N, P, Sigma, sigma, lambda)
}
\arguments{
\item{N}{Number of observations.}

\item{P}{Number of predictors.}

\item{Sigma}{Covariance matrix for predictors.}

\item{sigma}{Standard deviation of the noise term.}

\item{lambda}{Numeric vector of weights for different blocks.}
}
\value{
A data frame with predictors (X), outcome (Y), and true mean (mu).
}
\description{
Generates a dataset with correlated predictors and a nonlinear outcome structure.
}
\examples{
set.seed(123)
Sigma <- diag(20)
lambda <- c(0.25, 0.5, 0.75, 1)
sim_lfried(100, 20, Sigma, 10, lambda)

}
